refactor: переименовал компонент NavPage в Navigation
