+++ movies-explorer-frontend/src/components/Movies/Movies.tsx
@@ -1,6 +1,6 @@
 import './Movies.css'
 import Header from '../Header/Header';
-import NavPage from '../NavPage/NavPage';
+import Navigation from '../Navigation/Navigation';
 import SearchForm from '../SearchForm/SearchForm';
 import MoviesCardList from '../MoviesCardList/MoviesCardList';
 import Footer from '../Footer/Footer';
@@ -8,7 +8,7 @@ import Footer from '../Footer/Footer';
 function Movies() {
   return (
     <>
-      <Header children={<NavPage />} bgcolor="grey"/>
+      <Header children={<Navigation />} bgcolor="grey"/>
       <main className="movies">
         <SearchForm />
         <MoviesCardList />
