+++ movies-explorer-frontend/src/components/Navigation/Navigation.css
@@ -1,8 +1,8 @@
-.nav-page {
+.navigation {
   display: flex;
 }
 
-.nav-page__link {
+.navigation__link {
   font-size: 14px;
   font-weight: 400;
   line-height: 18px;
@@ -11,15 +11,15 @@
   text-decoration: none;
 }
 
-.nav-page__link:first-of-type {
+.navigation__link:first-of-type {
   display: none;
 }
 
-.nav-page__link_active {
+.navigation__link_active {
   font-weight: 500;
 }
 
-.nav-page__btn-account {
+.navigation__btn-account {
   font-size: 14px;
   font-weight: 500;
   line-height: 16px;
@@ -35,22 +35,22 @@
   background-size: 12px 14px;
 }
 
-.nav-page__icon-burger {
+.navigation__icon-burger {
   display: none;
 }
 
-.nav-page__container {
+.navigation__container {
   display: flex;
   flex-direction: row;
   align-items: center;
 }
 
-.nav-page__icon-close {
+.navigation__icon-close {
   display: none;
 }
 
 @media (max-width: 1023px) {
-  .nav-page__icon-burger {
+  .navigation__icon-burger {
     display: flex;
     width: 44px;
     height: 44px;
@@ -64,11 +64,11 @@
     background-color: inherit;
   }
 
-  .nav-page__icon-burger:hover {
+  .navigation__icon-burger:hover {
     cursor: pointer;
   }
 
-  .nav-page__overlay {
+  .navigation__overlay {
     display: none;
     position: fixed;
     top: 0;
@@ -79,7 +79,7 @@
     z-index: 2;
   }
 
-  .nav-page__container {
+  .navigation__container {
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -94,7 +94,7 @@
     box-sizing: border-box;
   }
 
-  .nav-page__icon-close {
+  .navigation__icon-close {
     display: flex;
     width: 76px;
     height: 76px;
@@ -109,53 +109,53 @@
     background-color: inherit;
   }
 
-  .nav-page__icon-close:hover {
+  .navigation__icon-close:hover {
     cursor: pointer;
   }
 
-  .nav-page__btn-account {
+  .navigation__btn-account {
     margin-top: auto;
     margin-left: 0;
   }
 
-  .nav-page__link:first-of-type {
+  .navigation__link:first-of-type {
     display: flex;
   }
 
-  .nav-page__link {
+  .navigation__link {
     font-weight: 500;
     font-size: 18px;
     line-height: 34px;
     margin: 0 0 16px;
   }
 
-  .nav-page__link_active {
+  .navigation__link_active {
     border-bottom: 2px solid #fff;
   }
 
-  .nav-page__overlay_display_flex {
+  .navigation__overlay_display_flex {
     display: flex;
   }
 }
 
 @media (max-width: 767px) {
-  .nav-page__icon-burger {
+  .navigation__icon-burger {
     width: 40px;
     height: 43px;
     background-size: 22px;
   }
 
-  .nav-page__container {
+  .navigation__container {
     left: 0;
   }
 
-  .nav-page__icon-close {
+  .navigation__icon-close {
     width: 67px;
     height: 67px;
     margin-bottom: 105px;
   }
 
-  .nav-page__link {
+  .navigation__link {
     line-height: 40px;
     margin-bottom: 6px;
   }
